Modifying products rest stuff

+++ src/main/java/com/gcu/data/DataAccessInterface.java
@@ -2,9 +2,11 @@ package com.gcu.data;
 
 import java.util.List;
 
+import com.gcu.model.ProductModel;
+
 public interface DataAccessInterface <T> {
 	public List<T> findAll();
-	public T findById(int id);
+	public T findById(long id);
 	public boolean create(T t);
 	public boolean update(T t);
 	public boolean delete(T t);
